requirements.txt
scratchattach~=2.0.9
discord~=1.7.3
discord.py~=2.4.0
python-dotenv~=1.0.1
google-generativeai~=0.8.3